+++ CHANGELOG.md
@@ -4,7 +4,6 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-
 ## [21.4.0] (unreleased)
 
 ### Added
@@ -16,16 +15,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 [21.4.0]: https://github.com/greenbone/ospd/compare/ospd-20.08...master
 
-## [20.8.2] (unreleased)
+## (20.8.2) - 2021-02-01
 
 ### Added
 - Allow the scanner to update total count of hosts. [#332](https://github.com/greenbone/ospd/pull/332)
+- Add more debug logging. [#352](https://github.com/greenbone/ospd/pull/352)
+- Set end_time for interrupted scans. [#353](https://github.com/greenbone/ospd/pull/353)
 
 ### Fixed
 - Fix OSP version. [#326](https://github.com/greenbone/ospd/pull/326)
 - Use empty string instead of None for credential. [#335](https://github.com/greenbone/ospd/pull/335)
 - Fix target_to_ipv4_short(). [#338](https://github.com/greenbone/ospd/pull/338)
 - Fix malformed target. [#341](https://github.com/greenbone/ospd/pull/341)
+- Initialize end_time with create_scan. [#354](https://github.com/greenbone/ospd/pull/354)
+- Fix get_count_total(). Accept -1 value set by the server. [#355](https://github.com/greenbone/ospd/pull/355)
 
 [20.8.2]: https://github.com/greenbone/ospd/compare/v20.8.1...ospd-20.08
 
@@ -98,7 +101,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 [20.8.0]: https://github.com/greenbone/ospd/compare/ospd-2.0...ospd-20.08
 
-## [2.0.1] (unreleased)
+## [2.0.1]
 
 ### Added
 - Add clean_forgotten_scans(). [#171](https://github.com/greenbone/ospd/pull/171)